backend/tests/test_smoke.py
"""Smoke tests for infrastructure validation."""

from django.db import connection

import pytest


class TestSmokeTests:
    """Smoke tests for pytest and Django."""

    def test_pytest_is_working(self):
        """Verify pytest works correctly."""
        assert True

    def test_django_settings_loaded(self, settings):
        """Verify Django settings are loaded."""
        assert settings is not None
        assert hasattr(settings, "DEBUG")

    def test_database_connection(self, db):
        """Verify database connection works."""
        with connection.cursor() as cursor:
            cursor.execute("SELECT 1")
            result = cursor.fetchone()
            assert result == (1,)

    def test_api_client_fixture(self, api_client):
        """Verify api_client fixture works."""
        assert api_client is not None
        assert hasattr(api_client, "get")
        assert hasattr(api_client, "post")

    def test_user_data_fixture(self, user_data):
        """Verify user_data fixture returns correct data."""
        assert "username" in user_data
        assert "email" in user_data
        assert "password" in user_data
        assert user_data["username"] == "testuser"

    def test_tutor_data_fixture(self, tutor_data):
        """Verify tutor_data fixture returns correct data."""
        assert "headline" in tutor_data
        assert "bio" in tutor_data
        assert "hourly_rate" in tutor_data
        assert "subjects" in tutor_data
        assert isinstance(tutor_data["subjects"], list)


class TestPytestMarkers:
    """Tests for custom pytest markers."""

    @pytest.mark.slow
    def test_slow_marker(self):
        """Verify slow marker works."""
        assert True

    @pytest.mark.integration
    def test_integration_marker(self):
        """Verify integration marker works."""
        assert True
