backend/requirements/dev.txt
-r base.txt

# Testing
pytest>=8.3,<9.0
pytest-django>=4.9,<5.0
pytest-cov>=6.0,<7.0
pytest-asyncio>=0.24,<1.0
factory-boy>=3.3,<4.0
faker>=33.1,<34.0

# Code Quality
ruff>=0.8,<1.0
pre-commit>=4.0,<5.0

# Type Checking
mypy>=1.13,<2.0
django-stubs>=5.1,<6.0
djangorestframework-stubs>=3.15,<4.0

# Development Utilities
ipython>=8.30,<9.0
django-extensions>=3.2,<4.0
Werkzeug>=3.1,<4.0
django-debug-toolbar>=4.4,<5.0

# Documentation
mkdocs>=1.6,<2.0
mkdocs-material>=9.5,<10.0
